3PMemoryCardController chipset, code cleanup, ...
there is also the emif further speed finetuning

understanding and working around a bit buggy memory card controller
chipset. Had to bring in a seperate memory card controller to help
make a product flexible and efficient with minimal complexity, at one
level, but at same time need to work with the chipset vendor to make
it less buggy and better in the long run.

Look at source and changelog as usual for some of the details